Move SparseVectorQueryBuilder and TextExpansionQueryBuilder to x-pack core (#117857)
This commit moves the SparseVectorQueryBuilder and TextExpansionQueryBuilder classes to the x-pack core module, enabling other modules to utilize these query builders.
Additionally, it introduces a SparseVectorQueryWrapper to extract sparse vector queries from standard Lucene queries.
This is needed for supporting semantic highlighting with sparse vector fields as follow up.
